5 games made in Python
A game - januar 2020

I have spend the month learning Python by building 5 games – 4 small classic games and 1 online multiplayer game.

The games are:
PONG
The classic game is a two-dimensional sports game that simulates table tennis. The player controls an in-game paddle by moving it vertically across the left or right side of the screen. They can compete against another player controlling a second paddle on the opposing side. Players use the paddles to hit a ball back and forth. You get points when your opponent doesn´t return the ball. Player A uses the arrow keys UP and DOWN and player B uses the w and s buttons to move up and down.
 

Snake
 In the game of Snake, the player uses the arrow keys to move a "snake" around the board.  The snake starts at the center of the board and can move left, right, up and down and through the walls. The snake moves at a constant speed. As the snake finds food, it eats the food, and thereby grows larger.  The game ends when the snake moves into itself.  The goal is to make the snake as large as possible before that happens.
 

Connect4
To win Connect Four, all you have to do is connect four of your colored checker pieces in a row. This can be done horizontally, vertically or diagonally. Each player will drop in one checker piece at a time. The player uses the mouse or mousepad to play.
 

Tetris
Tetris has very simple rules: you can only move the pieces in specific ways; your game is over if your pieces reach the top of the screen; and you can only remove pieces from the screen by filling all the blank space in a line. You use the arrow key UP to turn the piece the way you want it (4 ways), arrow key LEFT to move it to the left and arrow key RIGHT to move it to the right. If you want to speed it up you can use the arrow key DOWN to set the piece much faster. 
 

Online Multigame (Rock, Paper, Scissors)
 

This is the way it is supposed to look like but I can´t get it to work so mine looks like this so far:

 

After finishing the tutorials for the 5 game I made  my own game called Dropping hearts
 

The player uses the arrow keys to move the girl (paddle) from left to right. She is going to catch as many hearts as possible. If a heart is not cauth she will loose a life. She starts with 3 lifes and after loosing 3 lifes she will die.
The game needs another version of  PyCharm to function so I have made a new repository for that game alone.